added clear architecture, added mvvm
