+++ app/src/main/java/com/drozdova/app/domain/repository/DataRepo.kt
@@ -1,0 +1,7 @@
+package com.drozdova.app.domain.repository
+
+import com.drozdova.app.presentation.model.FilialModel
+
+interface DataRepo {
+    fun getDataList(): List<FilialModel>
+}
