+++ app/src/main/java/com/drozdova/app/presentation/MainActivity.kt
@@ -1,8 +1,10 @@
-package com.drozdova.app
+package com.drozdova.app.presentation
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.drozdova.app.R
 import com.drozdova.app.databinding.ActivityMainBinding
+import com.drozdova.app.presentation.MainFragment
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
