+++ app/src/main/res/layout/fragment_main.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    tools:context=".MainFragment">
+    tools:context=".presentation.MainFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
